refactor: re-structure and clean-up unused files
