+++ packages/dapp/src/components/Navbar/Menu/ChainsMenu/ChainsMenu.tsx
@@ -2,7 +2,8 @@ import { Box } from '@mui/material';
 import CircularProgress from '@mui/material/CircularProgress';
 import { useTheme } from '@mui/material/styles';
 import { useTranslation } from 'react-i18next';
-import { SubMenuKeys, useGetChains } from '../../../../const';
+import { SubMenuKeys } from '../../../../const';
+import { useGetChains } from '../../../../hooks';
 import { useMenu } from '../../../../providers/MenuProvider';
 import MenuItemComponent from '../../MenuItemComponent';
 import { NavbarMenu } from '../../index';
