--- packages/dapp/src/const/MenuConfig/ConnectedMenu/index.ts
@@ -1,2 +1,0 @@
-export { default as useConnectedMenu } from './Connected-MenuItems';
-export { default as useGetChains } from './Connected-SubMenu-Chains';
