+++ packages/dapp/src/const/index.ts
@@ -1,5 +1,4 @@
-export * from './linkMap';
-export * from './MenuConfig/ConnectedMenu';
 export * from './MenuConfig/MainMenu';
 export * from './MenuConfig/WalletMenu';
+export * from './linkMap';
 export * from './subMenuKeys';
