+++ packages/dapp/src/hooks/index.ts
@@ -1,2 +1,3 @@
+export { useGetChains } from './useGetChains';
 export * from './useUserTracking/';
 export { useWidgetConfig } from './useWidgetConfig';
